Added more variations of algorithms for experiments to find optimal parameter values; adjustments to run-sims and boxplots scripts.
