+++ nav-scripts/algorithm1/tiles-obstacle-detection-40.py
