+++ nav-scripts/algorithm2/reg-stuck-move-array-error-fix/pre-fix.py
