+++ nav-scripts/algorithm2/reg-stuck-move-array-error-fix/regions-average-no-reg-nav-halt-pf.py
