+++ nav-scripts/algorithm2/reg-stuck-move-array-error-fix/regions-average.py
