+++ nav-scripts/algorithm2/stuck-condition-fix/regions-average.py
