+++ nav-scripts/algorithm3/reg-stuck-move-array-error-stuck-fix/regions-weighted-average-norm-obs-det-40-pf.py
