+++ nav-scripts/algorithm3/reg-stuck-move-array-error-stuck-fix/regions-weighted-average-select-pf.py
