+++ nav-scripts/algorithm3/tiles-obstacle-detection-31875.py
@@ -6,7 +6,6 @@ import sys
 import time
 import numpy as np
 import inotify.adapters
-from scipy.stats import norm
 
 def goal_reached(x_pos, z_pos, x_goal_pos, z_goal_pos):
 	if ((x_pos >= x_goal_pos-0.5) and (x_pos <= x_goal_pos+0.5) and (z_pos >= z_goal_pos-0.5) and (z_pos <= z_goal_pos+0.5)):
@@ -39,13 +38,6 @@ def borderMove(border_angle):
 	elif border_angle < 0:
 		move_arr.append("l_bor")
 		return "l"
-
-def weightedMat(rows, columns, weights):
-	weighted_mat = np.zeros((rows,columns))
-	for x in range(0,rows):
-		for y in range(0,columns):
-			weighted_mat[x][y] = weights[x]
-	return weighted_mat
 
 i = inotify.adapters.Inotify()
 i.add_watch('/home/matt-ip/Desktop/ForestGenerator-1.2/frames/')
@@ -93,33 +85,9 @@ dir_check_count = dir_check_limit
 
 move_count = 0
 
-avg_obs_det_thresh = 50
-
-least_obs_arr = []
-
-weight_scale_arr = ["normal distribution", "linear", "logarithmic"]
-weight_scale = weight_scale_arr[0]
-
-row_indices = np.arange(1,601,1)
-weights = []
-
-if weight_scale == "normal distribution":
-	pdf = norm.pdf(row_indices, loc=300.5, scale=60)
-	for k in pdf:
-		weights.append(k/max(pdf))
-
-elif weight_scale == "linear":
-	for l in row_indices:
-		weights.append(l/600)
-
-elif weight_scale == "logarithmic":
-	base = 2
-	for m in row_indices:
-		weights.append((base**(int(m)-1))/(base**599))
-	
-weighted_mat_lr = weightedMat(600, 266, weights)
-weighted_mat_mid = weightedMat(600, 268, weights)
-weighted_mat_half = weightedMat(600, 400, weights)
+max_depth_output = 20
+obs_det_distance = 2.5
+obs_det_thresh = obs_det_distance * (255/max_depth_output)
 
 while goal_reached(true_x_pos, true_z_pos, true_x_goal_pos, true_z_goal_pos) == False:
 
@@ -231,7 +199,6 @@ while goal_reached(true_x_pos, true_z_pos, true_x_goal_pos, true_z_goal_pos) == 
 				print(border_move)				
 
 			obs_arr.append("N/A")
-			least_obs_arr.append("N/A")
 			
 			sys.stdout.flush()
 			time.sleep(0.1)
@@ -249,7 +216,6 @@ while goal_reached(true_x_pos, true_z_pos, true_x_goal_pos, true_z_goal_pos) == 
 					sys.stdout.flush()
 					time.sleep(0.1)
 					obs_arr.append("N/A")
-					least_obs_arr.append("N/A")
 				
 				elif goal_angle < 0:
 					print("l")
@@ -257,105 +223,69 @@ while goal_reached(true_x_pos, true_z_pos, true_x_goal_pos, true_z_goal_pos) == 
 					sys.stdout.flush()
 					time.sleep(0.1)
 					obs_arr.append("N/A")
-					least_obs_arr.append("N/A")
 
 			if dir_check_count < dir_check_limit:
 
 				img = cv2.imread(latest_file, 0) # 0 params, for grey image
 				rows, cols = img.shape[:2]  # image height and width
 
-				left_reg = img[0:rows, 0:int(round((1/3)*cols)-1)]
-				right_reg = img[0:rows, int(round((2/3)*cols)+1):cols]
-				mid_reg = img[0:rows, int(round((1/3)*cols)-1):int(round((2/3)*cols)+1)]
-				
-				left_mean, right_mean, mid_mean = np.mean(left_reg), np.mean(right_reg), np.mean(mid_reg)
-				reg_mean_arr = [left_mean, mid_mean, right_mean]
+				y1 = 0
+				x1 = 0
+				M = rows//20 # image height divided by 20
+				N = cols//20 # image width divided by 20
 
-				least_obs = np.argmax(reg_mean_arr)
-				most_obs = np.argmin(reg_mean_arr)
+				obstacle = False
 
-				left_w_avg = np.average(left_reg, weights = weighted_mat_lr)
-				right_w_avg = np.average(right_reg, weights = weighted_mat_lr)
-				mid_w_avg = np.average(mid_reg, weights = weighted_mat_mid)
+				obs_left = 0
+				obs_right = 0
 
-				reg_w_avg_arr = [left_w_avg, mid_w_avg, right_w_avg]
-				least_obs_w = np.argmax(reg_w_avg_arr)
-				least_obs_arr.append(least_obs_w)
-				most_obs_w = np.argmin(reg_w_avg_arr)
-				
-				if reg_w_avg_arr[most_obs_w] <= avg_obs_det_thresh:
+				for y in range(0,int(0.6*rows),M):
+					for x in range(0,cols,N):
+						y1 = y + M
+						x1 = x + N
+						tile = img[y:y+M, x:x+N]
+
+						thresh = cv2.threshold(tile, obs_det_thresh, 255, cv2.THRESH_BINARY_INV)[1]
+						pixels = cv2.countNonZero(thresh)
+
+						if pixels == (M * N):
+							obstacle = True
+							if x < 0.5*cols:
+								obs_left +=1
+							else:
+								obs_right +=1
+
+				if obstacle == False:
+					obs_arr.append("False")
+					print("w")
+					move_arr.append("w_gen")
+					dir_check_count += 1
+
+				elif obstacle == True:
 
 					obs_arr.append("True")
 
-					left_half = img[0:rows, 0:int(0.5*cols)]
-					right_half = img[0:rows, int(0.5*cols):cols]
-
-					left_half_mean, right_half_mean = np.mean(left_half), np.mean(right_half)
-					
-					left_half_w_avg = np.average(left_half, weights = weighted_mat_half)
-					right_half_w_avg = np.average(right_half, weights = weighted_mat_half)
-
-					if left_half_w_avg < right_half_w_avg:
-						if (move_arr[move_count-1] == "j_obs") or (move_arr[move_count-1] == "j_reg"):
+					if obs_left > obs_right:
+						if (move_arr[move_count-1] == "j_obs"):
 							print("j")
 							move_arr.append("j_obs")
 						else:
 							print("l")
 							move_arr.append("l_obs")
 					else:
-						if (move_arr[move_count-1] == "l_obs") or (move_arr[move_count-1] == "l_reg"):
+						if (move_arr[move_count-1] == "l_obs"):
 							print("l")
 							move_arr.append("l_obs")
 						else:
 							print("j")
 							move_arr.append("j_obs")
-
-				else:
-					
-					obs_arr.append("False")
 				
-					if ((x_pos >= x_goal_pos-10) and (x_pos <= x_goal_pos+10) and (z_pos >= z_goal_pos-10) and (z_pos <= z_goal_pos+10)):
-						print("w")
-						move_arr.append("w_gen")
-						dir_check_count += 1
-					
-					else:
-
-						if least_obs_w == 1:
-							print("w")
-							move_arr.append("w_reg")
-							dir_check_count += 1
-						
-						elif least_obs_w == 0:
-							if (left_w_avg > mid_w_avg-5) and (left_w_avg < mid_w_avg+5):
-								print("w")
-								move_arr.append("w_reg")
-								dir_check_count += 1
-							elif (move_arr[move_count-1] == "l_obs") or (move_arr[move_count-1] == "l_reg"):
-								print("l")
-								move_arr.append("l_reg")
-							else:
-								print("j")
-								move_arr.append("j_reg")
-						
-						elif least_obs_w == 2:
-							if (right_w_avg > mid_w_avg-5) and (right_w_avg < mid_w_avg+5):
-								print("w")
-								move_arr.append("w_reg")
-								dir_check_count += 1
-							elif (move_arr[move_count-1] == "j_obs") or (move_arr[move_count-1] == "j_reg"):
-								print("j")
-								move_arr.append("j_reg")
-							else:
-								print("l")
-								move_arr.append("l_reg")
-
 				sys.stdout.flush()
 				time.sleep(0.1)
-		
+
 		f6 = open("/home/matt-ip/Desktop/logs/debug.txt", "a")
 		if dir_check_count < dir_check_limit:
-			f6.write(str(move_count) + ": x: " + str(x_pos) + ", z: " + str(z_pos) + " -> Left w avg = " + str(left_w_avg) + ", Mid w avg = " + str(mid_w_avg) + ", Right w avg = " + str(right_w_avg) + " -> " + str(move_arr[move_count]) + "\n")
+			f6.write(str(move_count) + "-> " + str(move_arr[move_count]) + " : x: " + str(x_pos) + ", z: " + str(z_pos) + " -> obstacle = " + str(obstacle) + ", obs_left = " + str(obs_left) + ", obs_right = " + str(obs_right) + "\n")
 		else:
 			f6.write(str(move_count) + ": GOAL CORRECTION ACTIVE\n")
 		f6.close()
@@ -417,18 +347,14 @@ j_border_count = move_arr.count("j_bor")
 l_fixdir_count = move_arr.count("l_fixdir")
 j_fixdir_count = move_arr.count("j_fixdir")
 
-w_reg_count = move_arr.count("w_reg")
-l_reg_count = move_arr.count("l_reg")
-j_reg_count = move_arr.count("j_reg")
-
 w_gen_count = move_arr.count("w_gen")
 
 l_obs_count = move_arr.count("l_obs")
 j_obs_count = move_arr.count("j_obs")
 
-w_count = w_border_count + w_reg_count + w_gen_count
-l_count = l_border_count + l_fixdir_count + l_reg_count + l_obs_count
-j_count = j_border_count + j_fixdir_count + j_reg_count + j_obs_count
+w_count = w_border_count + w_gen_count
+l_count = l_border_count + l_fixdir_count + l_obs_count
+j_count = j_border_count + j_fixdir_count + j_obs_count
 
 euc_dis = (7/30) * w_count
 wobble_rate = (l_count + j_count) / move_count
