+++ run-sims.py
@@ -10,18 +10,19 @@ subprocess.run(["mv", "/home/matt-ip/Desktop/Testing123", "/home/matt-ip/Desktop
 #tree_density = "default"
 
 #algorithm_index_list = ["algorithm1", "algorithm2", "algorithm3"]
-algorithm_index_list = ["algorithm2","algorithm3"]
+algorithm_index_list = ["algorithm1"]
 
-sim_param = "density"
+#sim_param = "density"
+sim_param = "tiles-obstacle-detection-19125"
 #sim_param = "goal-correction"
 
-sim_param_variant_list = ["default", "sparse", "dense"]
-#sim_param_variant_list = ["default"]
+#sim_param_variant_list = ["default", "sparse", "dense"]
+sim_param_variant_list = ["sparse", "dense"]
 #sim_param_variant_list = ["goal-corr40"]
 
 seed_list = [1,2,3,5,6,7,8,9,10,11]
-#seed_list = [3,5,6,7,8,9,10,11]
-#seed_list = [1,2]
+#seed_list = [5,6,7,8,9,10,11]
+#seed_list = [1]
 
 #subprocess.run(["python3", "/home/matt-ip/Desktop/auto-forest-nav/nav-scripts/algorithm1/tiles-obstacle-detection.py", "|", "java", "-jar", "/home/matt-ip/Desktop/ForestGenerator-1.2/ForestGenerator-1.2-unix.jar", "/home/matt-ip/Desktop/ForestGenerator-1.2/src/main/resources/sparse.yaml", ">", "/home/matt-ip/Desktop/logs/cmdline-output-log.txt"])
 #os.system("python3 /home/matt-ip/Desktop/auto-forest-nav/nav-scripts/algorithm1/tiles-obstacle-detection.py | java -jar /home/matt-ip/Desktop/ForestGenerator-1.2/ForestGenerator-1.2-unix.jar /home/matt-ip/Desktop/ForestGenerator-1.2/src/main/resources/sparse.yaml > /home/matt-ip/Desktop/logs/cmdline-output-log.txt")
@@ -34,11 +35,11 @@ for algorithm in algorithm_index_list:
             #algorithm_name = "regions-average-no-reg-nav-halt"
             
             if algorithm == "algorithm1":
-                algorithm_name = "regions-weighted-average-select_log_base_1-1"
+                algorithm_name = "tiles-obstacle-detection-19125"
             elif algorithm == "algorithm2":
-                algorithm_name = "regions-average"
+                algorithm_name = "regions-average-obs-det-60"
             elif algorithm == "algorithm3":
-                algorithm_name = "regions-weighted-average"
+                algorithm_name = "tiles-obstacle-detection-31875"
             
             """
             if variant == "goal-corr40":
